chore: bump version to v5.0.7-rc.3
